Draw function

+++ src/libs/mouse.h
@@ -6,7 +6,7 @@
 
 #include "scene.h"
 
-extern float t_offset[2];
+extern double t_offset[2];
 
 void event_mouse_click(int button, int state, int x, int y);
 void event_mouse_motion(int x, int y);
